refactor matching and update gen-ai prompting
