+++ peerprep/backend/matching-service/src/routes/apiRoutes.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import { createRequest, deleteRequest, deleteSession } from '../controllers/requestController2';
+import { createRequest, deleteRequest, deleteSession } from '../controllers/requestController';
 import { matchStatusStream } from '../controllers/sseController'; // Import your SSE handler
 
 const router = express.Router();
